API1: moving header files into include directory
