+++ api1/include/redpitaya/rp1.h
@@ -1,0 +1,19 @@
+#ifndef RP1_H
+#define RP1_H
+
+#include "redpitaya/util.h"
+#include "redpitaya/uio.h"
+#include "redpitaya/evn.h"
+
+#include "redpitaya/hwid.h"
+
+#include "redpitaya/asg_bst.h"
+#include "redpitaya/asg_per.h"
+#include "redpitaya/gen_out.h"
+#include "redpitaya/gen.h"
+
+//#include "redpitaya/acq.h"
+//#include "redpitaya/osc.h"
+
+#endif
+
