+++ api1/meson.build
@@ -4,8 +4,8 @@ cc = meson.get_compiler('c')
 depm = cc.find_library('m', required : false)
 
 dep = [dependency('libudev'), depm]
-src = ['src/uio.c', 'src/evn.c', 'src/hwid.c', 'src/gen.c', 'src/asg_per.c', 'src/asg_bst.c', 'src/gen_out.c']
+src = ['src/util.c', 'src/uio.c', 'src/evn.c', 'src/hwid.c', 'src/gen.c', 'src/asg_per.c', 'src/asg_bst.c', 'src/gen_out.c']
 inc = include_directories(['include', 'src'])
 
 lib = shared_library('rp', src, dependencies : dep, include_directories : inc, version : '1.2.3', soversion : '0')
-executable('demo', 'src/main.c', dependencies : depm, link_with : lib)
+executable('demo', 'src/main.c', dependencies : depm, include_directories : inc, link_with : lib)
