+++ api1/src/asg_per.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <math.h>
 
-#include "util.h"
-#include "asg_per.h"
+#include "redpitaya/util.h"
+#include "redpitaya/asg_per.h"
 
 void rp_asg_per_init (rp_asg_per_t *handle, volatile rp_asg_per_regset_t *regset, double FS, int unsigned buffer_size, const fixp_t cnt_t) {
     handle->regset = regset;
