+++ api1/src/gen_out.c
@@ -2,8 +2,8 @@
 #include <stdbool.h>
 #include <math.h>
 
-#include "util.h"
-#include "gen_out.h"
+#include "redpitaya/util.h"
+#include "redpitaya/gen_out.h"
 
 void rp_gen_out_init(rp_gen_out_t *handle, volatile rp_gen_out_regset_t *regset, const fixp_t mul_t, const fixp_t sum_t) {
     handle->regset = regset;
