+++ api1/src/util.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-#include "util.h"
+#include "redpitaya/util.h"
 
 char * rp_util_fixp_print (const fixp_t value) {
     static char str[16];
